Update midarm 2 positioning

+++ brc_arm_motor_driver/brc_arm_motor_driver/roboclaw_3.py
@@ -769,13 +769,23 @@ class Roboclaw:
     def SetM1VelocityPID(self, address, p, i, d, qpps):
         # 		return self._write4444(address,self.Cmd.SETM1PID,long(d*65536),long(p*65536),long(i*65536),qpps)
         return self._write4444(
-            address, self.Cmd.SETM1PID, d * 65536, p * 65536, i * 65536, qpps
+            address,
+            self.Cmd.SETM1PID,
+            int(d * 65536),
+            int(p * 65536),
+            int(i * 65536),
+            int(qpps),
         )
 
     def SetM2VelocityPID(self, address, p, i, d, qpps):
         # 		return self._write4444(address,self.Cmd.SETM2PID,long(d*65536),long(p*65536),long(i*65536),qpps)
         return self._write4444(
-            address, self.Cmd.SETM2PID, d * 65536, p * 65536, i * 65536, qpps
+            address,
+            self.Cmd.SETM2PID,
+            int(d * 65536),
+            int(p * 65536),
+            int(i * 65536),
+            int(qpps),
         )
 
     def ReadISpeedM1(self, address):
